requirements.txt
blinker==1.4
Flask==0.10.1
Flask-LDAP==0.2.0
Flask-Mail==0.9.1
Flask-Script==2.0.5
Flask-WTF==0.12
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
msgpack-python==0.4.7
python-ldap==2.4.22
raven==5.8.1
redis==2.10.5
Werkzeug==0.11.2
wheel==0.24.0
WTForms==2.1
git+ssh://git@ghe.liwushuo.com/lws/flask-ldap.git@v0.7.0#egg=flask_ldap-0.7.0
